Makefile for generating output from MarkDown files
The makefile is based on http://git.io/GIs6og, and has been adapted to
also generate .docx files from .md input.
Additionally, it renames the output files to the last git commit hash,
so we can distinguish between the different files if needed.
Came in handy for a book chapter I'm writing...